Refactor according to the review
