+++ docs/development/getting-started.mdx
@@ -109,7 +109,6 @@ You can run Keep from your VSCode (after cloning the repo) by adding this config
             "PUSHER_APP_KEY": "keepappkey",
             "PUSHER_APP_SECRET": "keepappsecret",
             "LOG_FORMAT": "dev_terminal",
-            "DISABLE_POSTHOG": "true",
         }
       },
       {
