+++ keep-ui/.env.local.example
@@ -15,4 +15,3 @@ PUSHER_PORT=6001
 PUSHER_APP_KEY=keepappkey
 # Logging
 LOG_FORMAT=dev_terminal
-
